feat: add /health endpoint and degraded mode when DB is down
