+++ src/db/checkDatabase.ts
@@ -1,0 +1,10 @@
+import prisma from './prisma';
+
+export async function isDatabaseConnected(): Promise<boolean> {
+  try {
+    await prisma.$queryRaw`SELECT 1`;
+    return true;
+  } catch {
+    return false;
+  }
+}
